pgtbl partially moving to chal
