+++ src/platform/i386/Makefile
@@ -47,6 +47,7 @@ OBJS += miniacpi.o
 OBJS += vga.o
 OBJS += exception.o
 OBJS += lapic.o
+OBJS += chal_pgtbl.o
 
 COS_OBJ += pgtbl.o
 COS_OBJ += retype_tbl.o
